ENH: express sqrtm_psd in terms of a more general funm_psd
