+++ scipy/linalg/__init__.py
@@ -98,6 +98,7 @@ Matrix Functions
    sqrtm - Matrix square root
    sqrtm_psd - Matrix square root of a positive semi-definite matrix
    funm - Evaluating an arbitrary matrix function
+   funm_psd - Evaluating a matrix function of a positive semi-definite matrix
    expm_frechet - Frechet derivative of the matrix exponential
    expm_cond - Relative condition number of expm in the Frobenius norm
    fractional_matrix_power - Fractional matrix power
